# docs/operations/api_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portfolio Thesis Engine API",
    "description": "Backend service layer exposing PTE artefacts (canonical state, valuation, forecast, ficha, peers, cross-check logs) and the yaml authoring + pipeline-trigger workflow.",
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/tickers": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "List Tickers",
        "description": "All tickers with artefact-availability summary.",
        "operationId": "list_tickers_api_tickers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TickerSummary"
                  },
                  "type": "array",
                  "title": "Response List Tickers Api Tickers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/tickers/{ticker}": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Get Ticker",
        "operationId": "get_ticker_api_tickers__ticker__get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickers/{ticker}/canonical": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Get Canonical",
        "operationId": "get_canonical_api_tickers__ticker__canonical_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Canonical Api Tickers  Ticker  Canonical Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickers/{ticker}/valuation": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Get Valuation",
        "operationId": "get_valuation_api_tickers__ticker__valuation_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Valuation Api Tickers  Ticker  Valuation Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickers/{ticker}/ficha": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Get Ficha",
        "operationId": "get_ficha_api_tickers__ticker__ficha_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Ficha Api Tickers  Ticker  Ficha Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickers/{ticker}/forecast": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Get Forecast",
        "operationId": "get_forecast_api_tickers__ticker__forecast_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Forecast Api Tickers  Ticker  Forecast Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickers/{ticker}/peers": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Get Peers",
        "description": "Peers \u2014 combines yaml file + sqlite registry enrichment.",
        "operationId": "get_peers_api_tickers__ticker__peers_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Peers Api Tickers  Ticker  Peers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickers/{ticker}/cross-check": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Get Cross Check",
        "operationId": "get_cross_check_api_tickers__ticker__cross_check_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Cross Check Api Tickers  Ticker  Cross Check Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickers/{ticker}/pipeline-runs": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Get Pipeline Runs",
        "operationId": "get_pipeline_runs_api_tickers__ticker__pipeline_runs_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Pipeline Runs Api Tickers  Ticker  Pipeline Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TickerDetail": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "isin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isin"
          },
          "extraction_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Path"
          },
          "valuation_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valuation Path"
          },
          "forecast_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Path"
          },
          "ficha_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ficha Path"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "name",
          "profile",
          "currency",
          "exchange"
        ],
        "title": "TickerDetail",
        "description": "``GET /api/tickers/{ticker}``."
      },
      "TickerSummary": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "isin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isin"
          },
          "has_extraction": {
            "type": "boolean",
            "title": "Has Extraction",
            "default": false
          },
          "has_valuation": {
            "type": "boolean",
            "title": "Has Valuation",
            "default": false
          },
          "has_forecast": {
            "type": "boolean",
            "title": "Has Forecast",
            "default": false
          },
          "has_ficha": {
            "type": "boolean",
            "title": "Has Ficha",
            "default": false
          },
          "latest_extraction_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Extraction At"
          },
          "latest_valuation_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Valuation At"
          },
          "latest_forecast_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Forecast At"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "name",
          "profile",
          "currency",
          "exchange"
        ],
        "title": "TickerSummary",
        "description": "One row of ``GET /api/tickers``."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}